Support both sparse and dense vectors in fastembed

+++ swiftide-core/src/type_aliases.rs
@@ -2,7 +2,7 @@ pub type Embedding = Vec<f32>;
 pub type Embeddings = Vec<Embedding>;
 
 pub struct SparseEmbedding {
-    pub indices: Vec<u32>,
+    pub indices: Vec<usize>,
     pub values: Vec<f32>,
 }
 pub type SparseEmbeddings = Vec<SparseEmbedding>;
